bpf: don't perform revnat work on egress if not needed
We only really need it if we do know that we're dealing with a node port
connection, so we can skip it for all other traffic to avoid unnecessary
overhead of 2 CT map lookups in fast-path. Looking up is bad anyway given
CT actions e.g. on FIN have side-effects potentially for non-node-port
entries, therefore we should avoid these in general. Given the move into
tail call, we can also spare to attach to tc egress in general which removes
another layer we'd be hitting in fast-path, and it also might simplify
things for tunneling mode in future. CILIUM_CALL_IPV{4,6}_NODEPORT_REVNAT
will be in multiple call-maps but share the same maps (e.g. CT) otherwise.
Given we have already overloaded skb->mark heavily, this approach avoids
it completely; original idea would have been to use mark and then redirect
to tc egress.
