+++ bpf/lib/common.h
@@ -79,7 +79,9 @@
 #define CILIUM_CALL_IPV6_TO_ENDPOINT		14
 #define CILIUM_CALL_IPV4_NODEPORT_NAT		15
 #define CILIUM_CALL_IPV6_NODEPORT_NAT		16
-#define CILIUM_CALL_SIZE			17
+#define CILIUM_CALL_IPV4_NODEPORT_REVNAT	17
+#define CILIUM_CALL_IPV6_NODEPORT_REVNAT	18
+#define CILIUM_CALL_SIZE			19
 
 typedef __u64 mac_t;
